requirements.txt
keyboard
sounddevice
numpy
openai-whisper
requests
python-dotenv
